Move S3/GCS/HTTP readers to individual files
