+++ pkg/archive/archive_http.go
@@ -3,8 +3,11 @@ package archive
 import (
 	"archive/zip"
 	"context"
+	"io"
 	"net/http"
 	"slices"
+	"strconv"
+	"strings"
 
 	"github.com/pkg/errors"
 	"github.com/readium/go-toolkit/pkg/util/url"
@@ -92,5 +95,54 @@ func NewHTTPArchiveFactory(client *http.Client, config RemoteArchiveConfig) HTTP
 	return HTTPArchiveFactory{
 		client: client,
 		config: config,
+	}
+}
+
+// HTTP-specific reader
+type remoteHTTPReader struct {
+	client *http.Client
+	url    string
+	size   int64
+}
+
+func (r remoteHTTPReader) ReadRange(ctx context.Context, offset, length int64) (io.ReadCloser, error) {
+	if offset < 0 {
+		return nil, io.EOF
+	}
+
+	req, err := http.NewRequestWithContext(ctx, http.MethodGet, r.url, nil)
+	if err != nil {
+		return nil, err
+	}
+	var sb strings.Builder
+	sb.WriteString("bytes=")
+	sb.WriteString(strconv.FormatInt(offset, 10))
+	sb.WriteString("-")
+	if length > 0 {
+		sb.WriteString(strconv.FormatInt(offset+length-1, 10))
+	}
+	req.Header.Set("Range", sb.String())
+
+	resp, err := r.client.Do(req)
+	if err != nil {
+		return nil, err
+	}
+	if resp.StatusCode != http.StatusPartialContent {
+		return nil, errors.New("unexpected HTTP status code: " + strconv.Itoa(resp.StatusCode))
+	}
+
+	// User is responsible for closing the body
+	return resp.Body, nil
+}
+
+func (r remoteHTTPReader) Size() int64 {
+	return r.size
+}
+
+func RemoteArchiveReaderFromHTTP(client *http.Client, url url.AbsoluteURL, size int64) RemoteArchiveReader {
+	return &remoteHTTPReader{
+		client: client,
+		url:    url.String(),
+		size:   size,
 	}
 }
