+++ pkg/archive/archive_s3.go
@@ -3,7 +3,11 @@ package archive
 import (
 	"archive/zip"
 	"context"
+	"io"
+	"strconv"
+	"strings"
 
+	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 	"github.com/pkg/errors"
 	"github.com/readium/go-toolkit/pkg/util/url"
@@ -78,5 +82,49 @@ func NewS3ArchiveFactory(client *s3.Client, config RemoteArchiveConfig) S3Archiv
 	return S3ArchiveFactory{
 		client: client,
 		config: config,
+	}
+}
+
+// S3-specific reader
+type remoteS3Reader struct {
+	client *s3.Client
+	input  s3.GetObjectInput
+	head   s3.HeadObjectOutput
+}
+
+func (r remoteS3Reader) ReadRange(ctx context.Context, offset, length int64) (io.ReadCloser, error) {
+	if offset < 0 {
+		return nil, io.EOF
+	}
+
+	var sb strings.Builder
+	sb.WriteString("bytes=")
+	sb.WriteString(strconv.FormatInt(offset, 10))
+	sb.WriteString("-")
+	if length >= 0 {
+		sb.WriteString(strconv.FormatInt(offset+length-1, 10))
+	}
+	r.input.Range = aws.String(sb.String())
+	result, err := r.client.GetObject(ctx, &r.input)
+	if err != nil {
+		return nil, err
+	}
+
+	// User is responsible for closing the body
+	return result.Body, nil
+}
+
+func (r remoteS3Reader) Size() int64 {
+	if r.head.ContentLength != nil {
+		return *r.head.ContentLength
+	}
+	return 0
+}
+
+func RemoteArchiveReaderFromS3(client *s3.Client, output s3.HeadObjectOutput, input s3.GetObjectInput) RemoteArchiveReader {
+	return &remoteS3Reader{
+		client: client,
+		input:  input,
+		head:   output,
 	}
 }
